Adding block proposed and received metrics
