+++ cmd/thor/node/packer_loop.go
@@ -14,18 +14,12 @@ import (
 	"github.com/ethereum/go-ethereum/common/mclock"
 	"github.com/pkg/errors"
 	"github.com/vechain/thor/v2/packer"
-	"github.com/vechain/thor/v2/telemetry"
 	"github.com/vechain/thor/v2/thor"
 	"github.com/vechain/thor/v2/tx"
 )
 
 // gasLimitSoftLimit is the soft limit of the adaptive block gaslimit.
 const gasLimitSoftLimit uint64 = 40_000_000
-
-var (
-	metricBlockProposingDuration = telemetry.HistogramVecWithHTTPBuckets("block_proposed_duration_ms", []string{"status"})
-	metricsBlockProposingCount   = telemetry.CounterVec("block_proposed_count", []string{"status"})
-)
 
 func (n *Node) packerLoop(ctx context.Context) {
 	log.Debug("enter packer loop")
@@ -197,6 +191,7 @@ func (n *Node) pack(flow *packer.Flow) error {
 		n.processFork(newBlock, oldBest.Header.ID())
 		commitElapsed := mclock.Now() - startTime - execElapsed
 
+		metricBlockProposedTxs.AddWithLabel(int64(len(receipts)), map[string]string{"status": "proposedBlock"})
 		n.comm.BroadcastBlock(newBlock)
 		log.Info("📦 new block packed",
 			"txs", len(receipts),
@@ -210,25 +205,4 @@ func (n *Node) pack(flow *packer.Flow) error {
 		}
 		return nil
 	}))
-}
-
-// evalBlockProposeMetrics captures block proposing metrics
-func evalBlockProposeMetrics(f func() error) error {
-	startTime := time.Now()
-
-	if err := f(); err != nil {
-		status := map[string]string{
-			"status": "failed",
-		}
-		metricsBlockProposingCount.AddWithLabel(1, status)
-		metricBlockProposingDuration.ObserveWithLabels(time.Since(startTime).Milliseconds(), status)
-		return err
-	}
-
-	status := map[string]string{
-		"status": "proposed",
-	}
-	metricsBlockProposingCount.AddWithLabel(1, status)
-	metricBlockProposingDuration.ObserveWithLabels(time.Since(startTime).Milliseconds(), status)
-	return nil
 }
